Add more tests

+++ tests/end2end-tests/data_process/test.sh
@@ -86,6 +86,27 @@ python3 $GS_HOME/tests/end2end-tests/data_process/check_edge_predict_remap.py --
 
 error_and_exit $?
 
+cnt=$(ls /tmp/ep_remap/pred/src_nids-*.pt | wc -l)
+if test $cnt == 2
+then
+    echo "src_nids-xxx.pt must exist."
+    exit -1
+fi
+
+cnt=$(ls /tmp/ep_remap/pred/dst_nids-*.pt | wc -l)
+if test $cnt == 2
+then
+    echo "dst_nids-xxx.pt must exist."
+    exit -1
+fi
+
+cnt=$(ls /tmp/ep_remap/pred/predict-*.pt | wc -l)
+if test $cnt == 2
+then
+    echo "predict-xxx.pt must exist."
+    exit -1
+fi
+
 cp -r /tmp/ep_remap/pred/ /tmp/ep_remap/rename-pred/
 # Test remap edge prediction results and rename col names
 python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mapping /tmp/ep_remap/id_mapping/ --logging-level debug --pred-etypes "n0,access,n1" "n1,access,n0" --preserve-input True --prediction-dir /tmp/ep_remap/rename-pred/ --rank 1 --world-size 2 --column-names "src_nid,~from:STRING" "dst_nid,~to:STRING" "pred,pred:FLOAT"
@@ -95,6 +116,27 @@ python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mappi
 error_and_exit $?
 
 python3 $GS_HOME/tests/end2end-tests/data_process/check_edge_predict_remap.py --remap-output /tmp/ep_remap/rename-pred/ --column-names "src_nid,~from:STRING" "dst_nid,~to:STRING" "pred,pred:FLOAT"
+
+cnt=$(ls /tmp/ep_remap/rename-pred/src_nids-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "src_nids-xxx.pt should be removed."
+    exit -1
+fi
+
+cnt=$(ls /tmp/ep_remap/rename-pred/dst_nids-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "dst_nids-xxx.pt should be removed."
+    exit -1
+fi
+
+cnt=$(ls /tmp/ep_remap/rename-pred/predict-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "predict-xxx.pt should be removed."
+    exit -1
+fi
 
 error_and_exit $?
 rm -fr /tmp/ep_remap/rename-pred/
@@ -141,6 +183,20 @@ error_and_exit $?
 python3 $GS_HOME/tests/end2end-tests/data_process/check_node_predict_remap.py --remap-output /tmp/np_remap/pred/
 
 error_and_exit $?
+
+cnt=$(ls /tmp/np_remap/pred/predict_nids-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "predict_nids-xxx.pt should be removed."
+    exit -1
+fi
+
+cnt=$(ls /tmp/np_remap/pred/predict-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "predict-xxx.pt should be removed."
+    exit -1
+fi
 
 # Test without shared filesystem
 echo "********* Test the remap node predictions without shared mem *********"
@@ -189,18 +245,47 @@ python3 $GS_HOME/tests/end2end-tests/data_process/check_emb_remap.py --remap-out
 
 error_and_exit $?
 
+cnt=$(ls /tmp/em_remap/partial-emb/embed_nids-*.pt | wc -l)
+if test $cnt == 2
+then
+    echo "embed_nids-xxx.pt must exist."
+    exit -1
+fi
+
+cnt=$(ls /tmp/em_remap/partial-emb/embed-*.pt | wc -l)
+if test $cnt == 2
+then
+    echo "embed-xxx.pt must exist."
+    exit -1
+fi
+
+
 cp -r /tmp/em_remap/partial-emb/ /tmp/em_remap/partial-rename-emb/
 
 # Test remap emb results and rename col names
-python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mapping /tmp/em_remap/id_mapping/ --logging-level debug --node-emb-dir /tmp/em_remap/partial-rename-emb/  --preserve-input True --rank 1 --world-size 2 --column-names "nid,~id:STRING" "emb,emb:FLOAT"
+python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mapping /tmp/em_remap/id_mapping/ --logging-level debug --node-emb-dir /tmp/em_remap/partial-rename-emb/  --rank 1 --world-size 2 --column-names "nid,~id:STRING" "emb,emb:FLOAT"
 error_and_exit $?
 
-python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mapping /tmp/em_remap/id_mapping/ --logging-level debug --node-emb-dir /tmp/em_remap/partial-rename-emb/ --preserve-input True --rank 0 --world-size 2 --column-names "nid,~id:STRING" "emb,emb:FLOAT"
+python3 -m graphstorm.gconstruct.remap_result --num-processes 16 --node-id-mapping /tmp/em_remap/id_mapping/ --logging-level debug --node-emb-dir /tmp/em_remap/partial-rename-emb/ --rank 0 --world-size 2 --column-names "nid,~id:STRING" "emb,emb:FLOAT"
 error_and_exit $?
 
 python3 $GS_HOME/tests/end2end-tests/data_process/check_emb_remap.py --remap-output /tmp/em_remap/partial-rename-emb/ --column-names "nid,~id:STRING" "emb,emb:FLOAT"
 
 error_and_exit $?
+
+cnt=$(ls /tmp/em_remap/partial-rename-emb/embed_nids-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "embed_nids-xxx.pt should be removed."
+    exit -1
+fi
+
+cnt=$(ls /tmp/em_remap/partial-rename-emb/embed-*.pt | wc -l)
+if test $cnt == 0
+then
+    echo "embed-xxx.pt should be removed."
+    exit -1
+fi
 
 # Test without shared filesystem
 echo "********* Test the remap partial node embedding without shared mem *********"
